Added the ability to register members to a CSV
Group members can now register themselves with /register command. Their data gets saved in a CSV file along with the pickle. Used pandas for this and added installation command to the Dockerfile.
Also made it impossible for two same usernames to exist in the CSV.
Also added functions to test all of the above with their respective commands and CommandHandlers.